Move map mode icon creation to image utilities
