+++ NewHorizons/Utility/Files/ImageUtilities.cs
@@ -1,9 +1,12 @@
 using NewHorizons.Builder.Props;
+using NewHorizons.External;
+using NewHorizons.External.Modules.VariableSize;
 using NewHorizons.Utility.OWML;
 using OWML.Common;
 using System;
 using System.Collections.Generic;
 using System.IO;
+using System.Linq;
 using UnityEngine;
 
 namespace NewHorizons.Utility.Files
@@ -91,8 +94,8 @@ namespace NewHorizons.Utility.Files
             DeleteTexture(key, texture);
         }
 
-        public static void DeleteTexture(string key, Texture2D texture) 
-        { 
+        public static void DeleteTexture(string key, Texture2D texture)
+        {
             if (_textureCache.ContainsKey(key))
             {
                 if (_textureCache[key] == texture)
@@ -247,7 +250,7 @@ namespace NewHorizons.Utility.Files
             return texture;
         }
 
-        public static Texture2D MakeOutline(Texture2D texture, Color color, int thickness)
+        private static Texture2D MakeOutline(Texture2D texture, Color color, int thickness)
         {
             var key = $"{texture.name} > outline {color} {thickness}";
             if (_textureCache.TryGetValue(key, out var existingTexture)) return (Texture2D)existingTexture;
@@ -368,7 +371,7 @@ namespace NewHorizons.Utility.Files
             var pixels = image.GetPixels();
             for (int i = 0; i < pixels.Length; i++)
             {
-                var amount = (i % image.width) / (float) image.width;
+                var amount = (i % image.width) / (float)image.width;
                 var lightTint = LerpColor(lightTintStart, lightTintEnd, amount);
                 var darkTint = LerpColor(darkTintStart, darkTintEnd, amount);
 
@@ -488,6 +491,120 @@ namespace NewHorizons.Utility.Files
             var sprite = Sprite.Create(texture, new Rect(0, 0, texture.width, texture.height), new Vector2(0.5f, 0.5f), 100, 0, SpriteMeshType.FullRect, Vector4.zero, false);
             sprite.name = texture.name;
             return sprite;
+        }
+
+        public static Texture2D GetCachedOutlineOrCreate(NewHorizonsBody body, Texture2D original, string originalPath)
+        {
+            if (string.IsNullOrEmpty(originalPath))
+            {
+                Texture2D defaultTexture = null;
+                if (body.Config.Star != null)
+                {
+                    defaultTexture = ImageUtilities.GetTexture(Main.Instance, "Assets/DefaultMapModeStarOutline.png");
+                }
+                else if (body.Config.Atmosphere != null)
+                {
+                    defaultTexture = ImageUtilities.GetTexture(Main.Instance, "Assets/DefaultMapModNoAtmoOutline.png");
+                }
+                else
+                {
+                    defaultTexture = ImageUtilities.GetTexture(Main.Instance, "Assets/DefaultMapModePlanetOutline.png");
+                }
+
+                return defaultTexture;
+            }
+            else
+            {
+                var cachedPath = Path.Combine(body.Mod.ModHelper.Manifest.ModFolderPath, $"TextureCache_{Main.Instance.CurrentStarSystem}", originalPath);
+                var outlineTexture = ImageUtilities.GetTexture(body.Mod, cachedPath);
+
+                if (outlineTexture == null)
+                {
+                    NHLogger.LogVerbose($"Caching outline to {cachedPath}");
+
+                    var newTexture = ImageUtilities.MakeOutline(original, Color.white, 10);
+
+                    Directory.CreateDirectory(Path.GetDirectoryName(cachedPath));
+                    File.WriteAllBytes(cachedPath, newTexture.EncodeToPNG());
+
+                    return newTexture;
+                }
+                else
+                {
+                    return outlineTexture;
+                }
+            }
+        }
+
+        public static Texture2D AutoGenerateMapModePicture(NewHorizonsBody body)
+        {
+            Texture2D texture;
+
+            if (body.Config.Star != null)
+            {
+                texture = ImageUtilities.GetTexture(Main.Instance, "Assets/DefaultMapModeStar.png");
+            }
+            else if (body.Config.Atmosphere != null)
+            {
+                texture = ImageUtilities.GetTexture(Main.Instance, "Assets/DefaultMapModNoAtmo.png");
+            }
+            else
+            {
+                texture = ImageUtilities.GetTexture(Main.Instance, "Assets/DefaultMapModePlanet.png");
+            }
+
+            var color = GetDominantPlanetColor(body);
+            var darkColor = new Color(color.r / 3f, color.g / 3f, color.b / 3f);
+
+            texture = ImageUtilities.LerpGreyscaleImage(texture, color, darkColor);
+
+            return texture;
+        }
+
+        private static Color GetDominantPlanetColor(NewHorizonsBody body)
+        {
+            try
+            {
+                var starColor = body.Config?.Star?.tint;
+                if (starColor != null) return starColor.ToColor();
+
+                var atmoColor = body.Config.Atmosphere?.atmosphereTint;
+                if (body.Config.Atmosphere?.clouds != null && atmoColor != null) return atmoColor.ToColor();
+
+                if (body.Config?.HeightMap?.textureMap != null)
+                {
+                    try
+                    {
+                        var texture = ImageUtilities.GetTexture(body.Mod, body.Config.HeightMap.textureMap);
+                        var landColor = ImageUtilities.GetAverageColor(texture);
+                        if (landColor != null) return landColor;
+                    }
+                    catch (Exception) { }
+                }
+
+                var waterColor = body.Config.Water?.tint;
+                if (waterColor != null) return waterColor.ToColor();
+
+                var lavaColor = body.Config.Lava?.tint;
+                if (lavaColor != null) return lavaColor.ToColor();
+
+                var sandColor = body.Config.Sand?.tint;
+                if (sandColor != null) return sandColor.ToColor();
+
+                switch (body.Config?.Props?.singularities?.FirstOrDefault()?.type)
+                {
+                    case SingularityModule.SingularityType.BlackHole:
+                        return Color.black;
+                    case SingularityModule.SingularityType.WhiteHole:
+                        return Color.white;
+                }
+            }
+            catch (Exception)
+            {
+                NHLogger.LogWarning($"Something went wrong trying to pick the colour for {body.Config.name} but I'm too lazy to fix it.");
+            }
+
+            return Color.white;
         }
     }
 }
